add logs container
